// frontend/src/app/apiClient/model/casePatientAssessment.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BigDecimal } from 'bigdecimal';



export interface CasePatientAssessment { 
    patientAssessmentID: string;
    caseID: string;

    airwayUsingO2?: boolean;
    airwayPatientO2LPM?: BigDecimal;
    airwayUsingBVM?: boolean;
    airwayUsingETT?: boolean;
    airwayUsingETTSize?: BigDecimal;
    airwayUsingETTRate?: string;
    airwayUsingOPANPA?: boolean;
    airwayUsingTrach?: boolean;
    airwayUsingTrachSize?: BigDecimal;
    airwayNotes?: string;

    respBreathing?: Array<string>;
    respTracheaMidline?: boolean;
    respCough?: boolean;
    respCoughProductive?: string;
    respChestWallExpansion?: string;
    respChestTrauma?: boolean;
    respBreathSounds?: string;
    respBreathDiminishedRight?: boolean;
    respBreathDiminishedLeft?: boolean;
    respMonitors?: string;
    respSupplimentaryO2Device?: string;
    respEquipLPM?: BigDecimal;
    respEquipPercent?: BigDecimal;

    cardiacRate?: BigDecimal;
    cardiacRhythm?: string;
    cardiacSounds?: string;
    cardiacJvd?: string;
    cardiacPeripheralEdemaLocation?: string;
    cardiacPeripheralEdemaScore?: BigDecimal;
    cardiacExternalPacing?: boolean;
    cardiacExternalPacingMA?: BigDecimal;
    cardiacExternalPacingRate?: BigDecimal;
    cardiacEcgFindings?: string;
    cardiacNotes?: string;
    cardiacEquipment?: Array<string>;

    abdomenCondition?: Array<string>;
    abdomenTenderness?: Array<string>;
    abdomenBowelSounds?: string;
    abdomenFeedTube?: string;
    abdomenFeedTubeSize?: string;
    abdomenFeedTubeState?: string;
    abdomenNotes?: string;

    pelvisStable?: boolean;
    pelvisFoley?: boolean;
    pelvisFoleySize?: BigDecimal;
    pelvisAppearanceOfUrine?: string;
    pelvisNotes?: string;

    painDenies?: boolean;
    painLocation?: string;
    painSensation?: Array<string>;
    painScale?: BigDecimal;

    backTrauma?: Array<string>;
    backNotes?: string;

    extremitiesTrauma?: Array<string>;
    extremitiesPulsesRUE?: BigDecimal;
    extremitiesPulsesLUE?: BigDecimal;
    extremitiesPulsesRLE?: BigDecimal;
    extremitiesPulsesLLE?: BigDecimal;

    demeanorSpeech?: string;
    demeanorSkin?: Array<string>;
    demeanorBehavior?: string;

    neuroEyes?: string;
    neuroVerbal?: string;
    neuroMotor?: string;
    neuroGcsScore?: BigDecimal;
    neuroHeentTrauma?: boolean;
    neuroPupilsPERRLA?: boolean;
    neuroSizeLeft?: BigDecimal;
    neuroSizeRight?: BigDecimal;
    neuroReactionLeft?: BigDecimal;
    neuroReactionRight?: BigDecimal;
    neuroNotes?: string;

    diagramAnnotations?: string;
    diagramNotes?: string;

    overallSignature?: string;
}